pyproject.toml
[tool.poetry]
name = "introduction"
version = "0.1.0"
description = ""
authors = ["elzasimoes"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6.3"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
requests = "^2.31.0"
aiohttp = "^3.9.3"
mkdocs = "^1.5.3"
taskipy = "^1.12.2"
mkdocstrings = "^0.24.1"
mkdocs-material = "^9.5.13"
mkdocs-macros-plugin = "^1.0.5"
mkdocstrings-python = "^1.8.0"
black = "^24.2.0"
isort = "^5.13.2"
python-dotenv = "^1.0.1"

[tool.taskipy.tasks]
run = "uvicorn converter_api.main:app --reload --port 5000"
docs = "mkdocs serve"
lint = "isort . && black ."


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
